Combine spritesheets and json to minimize requests
**Before** we were downloading one png and one json metadata file for every animation for every character, making a total of 554 requests (15.5M) when loading the dance-party playground; not because we thought this was the _right_ thing to do, but because we were holding off on this optimization until our set of animations was fairly settled.

**After** we download one json file containing all the animation metadata, followed by five spritesheets per character, making a total of 60 requests (15.3M) when loading the dance-party playground.  I'm seeing local nocache playground load times consistently around 1200ms with the S3 cache configuration set properly and the cache warmed, where before they were closer to 2500.

A typical spritesheet is now ~2000x1200px, ~200k.  We're packing up to four long dances into a spritesheet, or up to seven short dances.  See https://github.com/code-dot-org/dance-spritesheets for details on this process, which uses `mogrify` to pre-resize our frames, `TexturePacker` to arrange them, and `pngquant` to make them as small as possible.

The generated metadata is structured for ease of use, with all costume and move names normalized to lowercase:

```json
{
  "costume": {
    "movename": {
      "spritesheet": "basename of the corresponding png file",
      "frames": [
        {
          frame: {},
          sourceSize: {},
          spriteSourceSize: {}
        }
      ]
    }
  }
}
```

This change depends on assets which I've already uploaded to a particular location on S3.