Enforce maximum event rates at API layer
Maximum of 10 events per measure -or- 10 events per second.

One or the other of these will be faster depending on the tempo of the song being played, but at the fastest tempo for any of our current songs (169bpm) we'll never fire an event more than once every 100ms, which should be a reasonable bound for performance reasons.