New effect commands with 'rand' support
Part of the fix for https://github.com/code-dot-org/dance-party/issues/469

Introduces `setBackgroundEffectWithPalette` and `setForegroundEffectExtended` which are pass-throughs to the existing functions, but exist so that we can create a new block instead of updating the old one for backwards-compatability.

Adds a random effect key 'rand' that, if passed, will cause a random effect to be selected from a preselected list appropriate to background/foreground as appropriate.